Moved code to the scope.js module
